chore: add routes utils and expose new stuff
